PHP:  - Gestion de la rotation des images (horaire et antihoraire)  - Suppression des images (fichier et DB)  - Mise a jour des infos 'Title' et 'Categories' sur le formulaire 'image' Javascript:  - Fullsize/Halfsize sur la page 'image'  - Gestion des misas a jour des infos (Avec Style)  - Gestion de la suppression/rotations image HTML/CSS:  - Gestion des resultats des mises a jour des infos image  - Modification des styles input/select (pour la mise a jour des infos)
git-svn-id: https://svn.knet-solutions.com/imgt4z@12 acda1053-8331-40ba-8dc0-498b8b23f828

+++ inc/class/ImagesManager.class.php
@@ -15,7 +15,20 @@ class ImagesManager{
 	}
 	
 	public function delete(Image $image){
-		$this->_db->exec("DELETE FROM images WHERE id = ". $image->id());
+		if(unlink('../storage/'.$image->userid().'/'.$image->timestamp().'.jpg')){
+			$this->_db->exec("DELETE FROM images WHERE id = ". $image->id());
+			return 1;
+		}else{
+			return 0;
+		}
+	}
+	
+	public function count($userid){
+		$q = $this->_db->prepare("SELECT COUNT(id) as count FROM images WHERE userid = :userid");
+		$q->bindValue(':userid',$userid);
+		$q->execute();
+		
+		return $q->fetch(PDO::FETCH_ASSOC);
 	}
 	
 	public function get($id){
@@ -39,11 +52,43 @@ class ImagesManager{
 		return $images;
 	}
 	
+	public function getListPaginated($userid,$start=0){
+		$limit = 10;
+		$images = array();
+		$q = $this->_db->prepare("
+			SELECT images.id as id, timestamp, orientation, categorieid, categories.name as categorie, userid, title, dateadd 
+			FROM images Inner Join categories ON images.categorieid = categories.id 
+			WHERE userid = :userid 
+			ORDER BY dateadd DESC 
+			LIMIT $start, $limit");
+		$q->bindValue(':userid',$userid);
+		/*$q->bindValue(':start',$start);
+		$q->bindValue(':limit',$limit);*/
+		$q->execute();
+		
+		while($datas = $q->fetch(PDO::FETCH_ASSOC)){
+			$images[] = new Image($datas);
+		}
+		return $images;
+	}
+	
 	public function update(Image $image){
 		$q = $this->_db->prepare("UPDATE images SET categorieid = :categorieid, title = :title WHERE id = :id");
 		$q->bindValue(':categorieid', $image->categorieid(), PDO::PARAM_INT);
 		$q->bindValue(':title', $image->title(), PDO::PARAM_STR);
-		$q->execute();
+		$q->bindValue(':id', $image->id(), PDO::PARAM_INT);
+		return $q->execute();
+	}
+	
+	public function updateOrientation(Image $image){
+		if($image->orientation() == 1){ $orientation = 0; }else{ $orientation = 1; }
+		$q = $this->_db->prepare("UPDATE images SET orientation = :orientation WHERE id = :id");
+		$q->bindValue(":orientation",$orientation, PDO::PARAM_INT);
+		$q->bindValue(":id",$image->id(),PDO::PARAM_INT);
+		if($q->execute()){
+			return $orientation;
+		}
+		
 	}
 	
 	public function setDb(PDO $db){
